only consider `-` part of word in lisp
Before, smartscan would consider `-` like any other letter.
This meant if you had code like:

    foo->next = NULL;
    free(foo);

and you try to use smartscan on `foo->next`, you'll get

    Forward scan for symbol "foo-"

...failing to reach the second instance of `foo`. This meant smartscan
would sometimes not give the right result in C and C++ code, which was a
real pain.

This commit fixes that problem by only including `-` in the syntax table
if the major mode is a lisp-like language.