Do not message the user on movement
As you wrote in your excellent [blog post on
eldoc](https://www.masteringemacs.org/article/seamlessly-merge-multiple-documentation-sources-eldoc),
space in the echo area is a precious resource. Therefore, I think movement commands should
not spam it with messages when they are doing exactly what they should. It also falls in
line with the unix rule of silence that programs should not say anything when everything
went as it should.